Initial CH341 Change
